test(discord): split thread chunk regressions
